refactor to take components out of pages directory, add absolute pathing for src folder for easy imports
