+++ pages/_app.tsx
@@ -1,7 +1,7 @@
-import '../styles/globals.css'
+import 'styles/globals.css'
 import type { AppProps } from 'next/app'
 import React from 'react';
-import { AppContext, GlobalContext } from './_context';
+import { AppContext, GlobalContext } from 'utils/context';
 function MyApp({ Component, pageProps }: AppProps) {
   return (
       <AppContext.Provider value={ GlobalContext }>
