+++ pages/signUp.tsx
@@ -1,4 +1,4 @@
-import AuthForm from './components/authForm'
+import AuthForm from 'components/authForm'
 import React from 'react';
 import { useRouter } from 'next/router';
 import axios from 'axios';
