+++ src/styles/AuthPage.module.scss
